- ParticleFadeOutAffector::setFadeOutTime can no longer be set to invalid values  - ParticleFadeOutAffector uses now throughout u32 for fadeOutTime (found by greenya)
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@3439 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/IParticleFadeOutAffector.h
@@ -20,14 +20,14 @@ public:
 	//! Sets the targetColor, i.e. the color the particles will interpolate to over time.
 	virtual void setTargetColor( const video::SColor& targetColor ) = 0;
 
-	//! Sets the amount of time it takes for each particle to fade out.
-	virtual void setFadeOutTime( f32 fadeOutTime ) = 0;
+	//! Sets the time in milliseconds it takes for each particle to fade out (minimal 1 ms)
+	virtual void setFadeOutTime( u32 fadeOutTime ) = 0;
 
 	//! Gets the targetColor, i.e. the color the particles will interpolate to over time.
 	virtual const video::SColor& getTargetColor() const = 0;
 
-	//! Gets the amount of time it takes for each particle to fade out.
-	virtual f32 getFadeOutTime() const = 0;
+	//! Gets the time in milliseconds it takes for each particle to fade out.
+	virtual u32 getFadeOutTime() const = 0;
 
 	//! Get emitter type
 	virtual E_PARTICLE_AFFECTOR_TYPE getType() const { return EPAT_FADE_OUT; }
